# backend/docs/swagger.yaml
basePath: /
definitions:
...
  db.Quiz:
    properties:
      createdAt:
        type: string
      creatorID:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  db.QuizQuestion:
    properties:
      correctAns:
        type: string
      id:
        type: string
      options:
        items:
          type: integer
        type: array
      questionText:
        type: string
      questionType:
        $ref: '#/definitions/db.QuestionType'
      quizID:
        type: string
    type: object
  db.QuizSubmission:
    properties:
      answers:
        items:
          type: integer
        type: array
      id:
        type: string
      kaderID:
        type: string
      quizID:
        type: string
      score:
        type: number
      submittedAt:
        type: string
    type: object
...
  handler.CreateQuizQuestionRequest:
    properties:
      correct_ans:
        type: string
      options:
        items:
          type: integer
        type: array
      question_text:
        type: string
      question_type:
        $ref: '#/definitions/db.QuestionType'
    required:
    - options
    - question_text
    - question_type
    type: object
  handler.CreateQuizRequest:
    properties:
      description:
        type: string
      title:
        type: string
    required:
    - title
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
  handler.MessageResponse:
    properties:
      message:
        example: operation successful
        type: string
    type: object
...
  handler.UpdateQuizQuestionRequest:
    properties:
      correct_ans:
        type: string
      options:
        items:
          type: integer
        type: array
      question_text:
        type: string
      question_type:
        $ref: '#/definitions/db.QuestionType'
    required:
    - options
    - question_text
    - question_type
    type: object
  handler.UpdateQuizRequest:
    properties:
      description:
        type: string
      title:
        type: string
    required:
    - title
    type: object
...
info:
  contact: {}
  description: Backend REST API for Centing (Stunting Prevention & Monitoring System).
  title: Centing API
  version: "1.0"
paths:
...
  /kader/submissions:
    get:
      description: Retrieve all quiz submissions submitted by the authenticated kader
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.QuizSubmission'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: insufficient permissions'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List quiz submissions by kader (Kader only)
      tags:
      - Quizzes
...
  /nakes/quizzes:
    post:
      consumes:
      - application/json
      description: Create a new educational quiz
      parameters:
      - description: Create Quiz Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateQuizRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.Quiz'
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: insufficient permissions'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create quiz (Nakes only)
      tags:
      - Quizzes
  /nakes/quizzes/{id}:
    delete:
      description: Delete a quiz by ID
      parameters:
      - description: Quiz ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Invalid quiz ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: insufficient permissions'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete quiz (Nakes only)
      tags:
      - Quizzes
    put:
      consumes:
      - application/json
      description: Update an existing quiz title and description
      parameters:
      - description: Quiz ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Update Quiz Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateQuizRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.Quiz'
        "400":
          description: Validation error or invalid ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: insufficient permissions'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update quiz (Nakes only)
      tags:
      - Quizzes
  /nakes/quizzes/{id}/questions:
    post:
      consumes:
      - application/json
      description: Add a question with answer options to a quiz
      parameters:
      - description: Quiz ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Create Question Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.CreateQuizQuestionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/db.QuizQuestion'
        "400":
          description: Validation error or invalid quiz ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: insufficient permissions'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create quiz question (Nakes only)
      tags:
      - Quizzes
  /nakes/quizzes/{id}/questions/{question_id}:
    delete:
      description: Delete a question from a quiz
      parameters:
      - description: Quiz ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Question ID (UUID)
        in: path
        name: question_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Invalid question ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: insufficient permissions'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete quiz question (Nakes only)
      tags:
      - Quizzes
    get:
      description: Retrieve single quiz question details
      parameters:
      - description: Quiz ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Question ID (UUID)
        in: path
        name: question_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.QuizQuestion'
        "400":
          description: Invalid question ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: insufficient permissions'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get quiz question by ID (Nakes only)
      tags:
      - Quizzes
    put:
      consumes:
      - application/json
      description: Update question text, type, options, or correct answer
      parameters:
      - description: Quiz ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Question ID (UUID)
        in: path
        name: question_id
        required: true
        type: string
      - description: Update Question Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateQuizQuestionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.QuizQuestion'
        "400":
          description: Validation error or invalid ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: insufficient permissions'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update quiz question (Nakes only)
      tags:
      - Quizzes
  /nakes/quizzes/{id}/submissions:
    get:
      description: List all submissions for a quiz
      parameters:
      - description: Quiz ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.QuizSubmission'
            type: array
        "400":
          description: Invalid quiz ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: insufficient permissions'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List quiz submissions (Nakes only)
      tags:
      - Quizzes
  /nakes/quizzes/{id}/submissions/{submission_id}:
    delete:
      description: Delete a quiz submission by ID
      parameters:
      - description: Quiz ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Submission ID (UUID)
        in: path
        name: submission_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Invalid submission ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: insufficient permissions'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete quiz submission (Nakes only)
      tags:
      - Quizzes
    get:
      description: Retrieve single quiz submission details
      parameters:
      - description: Quiz ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Submission ID (UUID)
        in: path
        name: submission_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.QuizSubmission'
        "400":
          description: Invalid submission ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: 'Forbidden: insufficient permissions'
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get quiz submission by ID (Nakes only)
      tags:
      - Quizzes
...
  /quizzes:
    get:
      description: Retrieve paginated list of quizzes
      parameters:
      - description: Limit (default 10)
        in: query
        name: limit
        type: integer
      - description: Offset (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Quiz'
            type: array
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List quizzes
      tags:
      - Quizzes
  /quizzes/{id}:
    get:
      description: Retrieve quiz details by ID
      parameters:
      - description: Quiz ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/db.Quiz'
        "400":
          description: Invalid quiz ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get quiz by ID
      tags:
      - Quizzes
  /quizzes/{id}/questions:
    get:
      description: Retrieve all questions for a specific quiz
      parameters:
      - description: Quiz ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.QuizQuestion'
            type: array
        "400":
          description: Invalid quiz ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List quiz questions
      tags:
      - Quizzes
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"